+++ src/steps/account/index.ts
@@ -3,6 +3,7 @@ import {
   IntegrationStepExecutionContext,
 } from '@jupiterone/integration-sdk-core';
 
+import { createAPIClient } from '../../client';
 import { IntegrationConfig } from '../../config';
 import { Steps, Entities } from '../constants';
 import { createAccountEntity } from './converter';
@@ -11,8 +12,14 @@ export const ACCOUNT_ENTITY_KEY = 'entity:account';
 
 export async function fetchAccountDetails({
   jobState,
+  instance,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
-  const accountEntity = await jobState.addEntity(createAccountEntity());
+  const apiClient = createAPIClient(instance.config);
+
+  const currentUser = await apiClient.getCurrentUser();
+  const accountEntity = await jobState.addEntity(
+    createAccountEntity(currentUser),
+  );
 
   await jobState.setData(ACCOUNT_ENTITY_KEY, accountEntity);
 }