+++ src/steps/index.ts
@@ -1,6 +1,15 @@
 import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { organizationSteps } from './organization';
+import { organizationMemberSteps } from './organization-member';
+import { teamSteps } from './team';
+import { repositorySteps } from './repository';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [
+  ...accountSteps,
+  ...organizationSteps,
+  ...organizationMemberSteps,
+  ...teamSteps,
+  ...repositorySteps,
+];
 
 export { integrationSteps };